Cleanup competition code from test_state_publisher.py
